+++ tests/errors/index.test.js
@@ -4,14 +4,14 @@ const test = require('node:test');
 const assert = require('node:assert');
 const path = require('node:path');
 const Script = require('../..');
-const { require: read } = Script;
+const { from: read } = Script;
 const exec = Script.execute;
 
 const target = name => path.join(__dirname, 'examples', name);
 
 test('[CORE] Eval error ', async () => {
   try {
-    exec(`module.exports = eval('100 * 2');`, { type: 'cjs' });
+    exec(`module.exports = eval('100 * 2');`);
     assert.fail(new Error('Should throw an error.'));
   } catch (error) {
     assert.strictEqual(error.constructor.name, 'EvalError');